% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2.r
\docType{data}
\name{dat2}
\alias{dat2}
\title{Phenotypes and 2 sets of PGSs}
\format{
A data frame with 1000 rows and 3 variables:
\describe{
\item{V1}{Phenotype}
\item{V2}{PGS1, regulatory region}
\item{V3}{PGS2, non-regulatory region}

}
}
\usage{
dat2
}
\description{
A dataset containing phenotypes and 2 sets of PGSs estimated from 2 sets of SNPs from regulatroy and non-regulatory genomic regions
}
\keyword{datasets}
