% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoa1_2.r
\name{olkin1_2}
\alias{olkin1_2}
\title{olkin1_2 function}
\usage{
olkin1_2(omat, nv)
}
\arguments{
\item{omat}{3 by 3 matrix having the correlation coefficients between y, x1 and x2, i.e. omat=cor(dat) where dat is N by 3 matrix having variables in the order of cbind (y,x1,x2)}

\item{nv}{Sample size}
}
\value{
This function will be used as source code
}
\description{
olkin1_2 function
}
\keyword{source}
