% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc_trf.r
\name{cc_trf}
\alias{cc_trf}
\title{cc_trf function}
\usage{
cc_trf(R2, se, K, P)
}
\description{
This function transforms the predictive ability and
its standard error between the observed scale
and liability scale cc_trf(R2,se, K,P), where
R2 is the R squared value of the model for either
observed scale or liability scale, se is the
standard error for either observed scale or liability scale,
K and P is the prevalence of cases in the population.
}
\examples{
To get the transformed predictive ability 
output=cc_trf(0.06, 0.002, 0.05, 0.05)

output$R2l (transformed R2 to liability scale, assuming 0.06 in observed scale)
0.2679337

output$sel (transformed se to liability scale, assuming 0.002 is observed scale)
0.008931123

output$R2O (transformed R2 to observed scale, assuming 0.06 in liability scale)
0.01343616

output$seO (transformed to in observed scale, assuming 0.002 is liability scale)
0.000447872

}
\references{
Lee, S. H., Goddard, M. E., Wray, N. R., and Visscher, P. M. A better coefficient of determination for genetic profile analysis. Genetic epidemiology,(2012). 36(3): p. 214-224.
}
\keyword{R2}
\keyword{Transformation}
\keyword{and}
\keyword{between}
\keyword{liability}
\keyword{observed}
\keyword{of}
\keyword{scale}
