% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoa_beta1_2.r
\name{olkin_beta1_2}
\alias{olkin_beta1_2}
\title{olkin_beta1_2 function}
\usage{
olkin_beta1_2(omat, nv)
}
\arguments{
\item{omat}{3 by 3 matrix having the correlation coefficients between y, x1 and x2, i.e. omat=cor(dat) where dat is N by 3 matrix having variables in the order of cbind (y,x1,x2)}

\item{nv}{sample size}
}
\description{
This function derives Information matrix for beta1^2 and beta2^2
where beta1 and 2 are regression coefficients from a multiple regression model,
i.e. y = x1•beta1 + x2•beta2 + e, where y, x1 and x2 are column-standardised
(see Olkin and Finn 1995).
}
\examples{
olkin_beta1_2 (omat,nv) (see example file)
dat=read.table("test_ukbb_thresholds_scaled") (see example files)
omat=cor(dat)
omat
1.000 0.195 0.197 0.197 0.197 0.194 0.191 0.183 0.168 0.130 0.103
0.195 1.000 0.998 0.995 0.992 0.984 0.964 0.936 0.860 0.757 0.669
0.197 0.998 1.000 0.998 0.996 0.988 0.969 0.942 0.867 0.764 0.674
0.197 0.995 0.998 1.000 0.998 0.991 0.972 0.946 0.872 0.769 0.679
0.197 0.992 0.996 0.998 1.000 0.995 0.978 0.953 0.879 0.777 0.687
0.194 0.984 0.988 0.991 0.995 1.000 0.988 0.966 0.897 0.796 0.706
0.191 0.964 0.969 0.972 0.978 0.988 1.000 0.987 0.928 0.831 0.742
0.183 0.936 0.942 0.946 0.953 0.966 0.987 1.000 0.959 0.869 0.780
0.168 0.860 0.867 0.872 0.879 0.897 0.928 0.959 1.000 0.942 0.859
0.130 0.757 0.764 0.769 0.777 0.796 0.831 0.869 0.942 1.000 0.951
0.103 0.669 0.674 0.679 0.687 0.706 0.742 0.780 0.859 0.951 1.000
}
\references{
Olkin, I. and J.D. Finn, Correlations redux. Psychological Bulletin, 1995. 118(1): p. 155.
}
\keyword{context}
\keyword{correlation}
\keyword{in}
\keyword{information}
\keyword{matrix}
\keyword{of}
\keyword{the}
