% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoa_beta_inf.r
\name{olkin_beta_inf}
\alias{olkin_beta_inf}
\title{olkin_beta_inf function}
\usage{
olkin_beta_inf(omat, nv)
}
\arguments{
\item{omat}{3 by 3 matrix having the correlation coefficients between y, x1 and x2, i.e. omat=cor(dat) where dat is N by 3 matrix having variables in the order of cbind (y,x1,x2)}

\item{nv}{sample size}
}
\description{
This function derives Information matrix for beta1 and beta2
where beta1 and 2 are regression coefficients from a multiple regression model,
i.e. y = x1•beta1 + x2•beta2 + e, where y, x1 and x2 are column-standardised
(see Olkin and Finn 1995).
}
\examples{
olkin_beta_inf (omat,nv) (see example file)
dat=read.table("test_ukbb_thresholds_scaled") (see example files)
omat=cor(dat)[1:3,1:3]
omat
1.0000000 0.1958636 0.1970060
0.1958636 1.0000000 0.9981003
0.1970060 0.9981003 1.0000000 
}
\references{
Olkin, I. and J.D. Finn, Correlations redux. Psychological Bulletin, 1995. 118(1): p. 155.
}
\keyword{context}
\keyword{correlation}
\keyword{in}
\keyword{information}
\keyword{matrix}
\keyword{of}
\keyword{the}
