% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_diff2.r
\name{r2_diff}
\alias{r2_diff}
\title{r2_diff function}
\usage{
r2_diff(dat, v1, v2, nv)
}
\arguments{
\item{dat}{N by (M+1) matrix having variables in the order of cbind(y,x)}

\item{v1}{This can be set as v1=c(1) or v1=c(1,2)}

\item{v2}{This can be set as v2=c(2), v2=c(3), v2=c(1,3) or v2=c(3,4)}

\item{nv}{sample size}
}
\description{
This function estimates var(R2(y~x[,v1]) - R2(y~x[,v2]))
where R2 is the R squared value of the model,
y is N by 1 matrix having the dependent variable, and
x is N by M matrix having M explanatory variables.
v1 or v2 indicates the ith column in the x matrix
(v1 or v2 can be multiple values between 1 - M, see Arguments below)
}
\examples{
To get the test statistics for the difference between R2(y~x[,v1]) and R2(y~x[,v2]).
(here we define R2_1=R2(y~x[,v1])) and R2_2=R2(y~x[,v2])))

dat=read.table("test_ukbb_thresholds_scaled") (see example files)
nv=length(dat$V1)
v1=c(1)
v2=c(2)
output=r2_diff(dat,v1,v2,nv)
output

r2redux output

output$var1 (variance of R2_1)
0.0001437583

output$var2 (variance of R2_2)
0.0001452828

output$var_diff (variance of difference between R2_1 and R2_2)
5.678517e-07

output$r2_based_p (p-value for significant difference between R2_1 and R2_2)
0.5514562

output$mean_diff (differences between R2_1 and R2_2)
-0.0004488044

output$upper_diff (upper limit of 95\% CI for the difference)
0.001028172

output$lower_diff (lower limit of 95\% CI for the difference)
-0.001925781



To get the test statistics for the difference between R2(y~x[,v1]+x[,v2]) and R2(y~x[,v2]).
(here R2_1=R2(y~x[,v1]+x[,v2]) and R2_2=R2(y~x[,v1]))

dat=read.table("test_ukbb_thresholds_scaled") (see example files)
nv=length(dat$V1)
v1=c(1,2)
v2=c(1)
output=r2_diff(dat,v1,v2,nv)
output

r2redux output

output$var1 (variance of R2_1)
0.0001475195

output$var2 (variance of R2_2)
0.0001437583

output$var_diff (variance of difference between R2_1 and R2_2)
2.321425e-06

output$r2_based_p (p-value for significant difference between R2_1 and R2_2)
0.4369177

output$mean_diff (differences between R2_1 and R2_2)
0.0006042383

output$upper_diff (upper limit of 95\% CI for the difference)
0.004887989

output$lower_diff (lower limit of 95\% CI for the difference)
-0.0005574975


}
\keyword{R2}
\keyword{information}
\keyword{matrix}
\keyword{variance}
