% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2_var2.r
\name{r2_var}
\alias{r2_var}
\title{r2_var function}
\usage{
r2_var(dat, v1, nv)
}
\arguments{
\item{dat}{N by (M+1) matrix having variables in the order of cbind(y,x)}

\item{v1}{This can be set as v1=c(1), v1=c(1,2) or possibly with more values}

\item{nv}{sample size}
}
\description{
This function estimates var(R2(y~x[,v1]))
where R2 is the R squared value of the model,
where R2 is the R squared value of the model,
y is N by 1 matrix having the dependent variable, and
x is N by M matrix having M explanatory variables.
v1 indicates the ith column in the x matrix
(v1 can be multiple values between 1 - M, see Arguments below)
}
\examples{
To get the test statistics for R2(y~x[,v1])

dat=read.table("test_ukbb_thresholds_scaled") (see example file)
nv=length(dat$V1)
v1=c(1) 
output=r2_var(dat,v1,nv)

r2redux output

output$var (variance of R2) 
0.0001437583

output$r2_based_p (P-value under the null hypothesis, i.e. R2=0)
1.213645e-10

output$mean_r2 (R2)
0.03836254

output$upper_r2 (upper limit of 95\% CI for R2)
0.06435214

output$lower_r2 (lower limit of 95\% CI for R2)
0.01763347



To get the test statistic for R2(y~x[,v1]+x[,v2]+x[,v3])

dat=read.table("test_ukbb_thresholds_scaled") (see example file)
nv=length(dat$V1)
v1=c(1,2,3) ##1=R2(y~x[,v1]+x[,v2]+x[,v3])
outout=r2_var(dat,v1,nv)
output

r2redux output

output$var (variance of R2)
0.0001499374

output$r2_based_p (R2 based P-value)
7.461267e-11

output$mean_r2 (R2)
0.03917668

output$upper_r2 (upper limit of 95\% CI for R2)
0.06538839

output$lower_r2 (lower limit of 95\% CI for R2)
0.01821657
}
\keyword{R2}
\keyword{information}
\keyword{matrix}
\keyword{variance}
